Add series-specific destination
In addition to the default destination, save chapters to a directory
designated to the series. Upon retrieving the chapters, construct a path
using the series and the 'DefaultDestination' as a base.

Allow users to switch between the origial save destination, which is
separate from the 'DefaultDestination', and the series destination.

+ Replace the save destination TextBox with a Label control
- Remove the supported sites panel