Support single-page hosts
Short-circuit the default behavior of the application by listing the URIs
which host the chapter's images by allowing the 'single_page' value to be
set in the 'host'section.

Once done, the application will ignore the 'pages' section and any values
therein. Effectively, the first page will be the only page which will be
parsed.