Include enum for specifying language options
Instead of supplying an array with the words to be excluded, the
appropriate array may be selected based on the language requested.

Including Japanese and English causes problems based on the rules of each
language and needs to be separated. The 'ToTitleCase' method follows
English rules which aren't applicable to title-casing Japanese sentences
with Romaji. The English rules need to be enabled when English designated
sentences have been supplied.