Set chapter as float and add wildcard placeholders
A work-around for the string formatter rounding the chapters with decimal
places is to declare them as float. The value won't be rounded in that
case.

Adding a single placeholder (#) character with a value that has 2 decimal
places will cause it to be rounded. The current solution is to increase
the number of placeholder to 7 characters, the maximum precision of the
Float data type.