Fix MangaFox regression
A bug was introduced when moving the HostVariables to the base title and
chapter classes. In doing so, the MangaFox plugin broke when the
'address_trimmed' variable was called. It refers to the main page
instead of the chapter's page.

The output is 'fanfox.net/anne_freaks/1.html' versus the expected
'fanfox.net/anne_freaks/v01/c005/1.html'.

In addition, MangaPark's host configuration breaks with the correct
output.

Fix the MangaFox bug by moving the code which sets 'address' and
'address_trimmed' to the ChapterDistill class constructor. Prefix
'series_' to the incorrectly placed 'address' and 'address_trimmed' code
in the TitleDistill constructor.