Introduce retrying when downloading the chapters
If during the download of the chapter pages, the download fails, the
application starts downloading everything from the beginning.

To conserve the user's bandwidth and time, introduce a retry loop. Each
time the download fails, catche the exception and increment the retry
count by one. Once the limit has been reached, throw the exception.

Currently, the application retries an infinite amount of times, so this
could become an issue in of its own. Use this as a temporary solution
until the download service has been unified.