Hide formatting buttons in RELEASE
Due to the complexity of the formatting solution (renaming chapters and
pages), hide the button until safe-guards have been implemented.

A logging system should be built into the application to show the user
what was changed, what it was, and what it is. Due to the varying nature
of the websites, the chances of incorrect naming or other bugs occuring is
high. With the proposed logging system, the user or developer will be able
to return the chapters to their original form, should it be required.