Add generic serializer class
In order to unify the code and reduce the possibility of bugs, create a
generic serializer for the various manager classes.

Each manager is to be built atop the last. The main window setting
manager will have the least functionality; Save and Load. The bookmark
manager deals with a list of Bookmark-typed instances and will require
Add and Remove methods, among dealing with the serialization of the array.