Fix bug where '__inner_text' isn't honored
Fix the bug in GenericParseAction<T> where the attribute's value is
retrieved. The '__inner_text' meta-attribute is only retrieved in the
block following the non-empty regex post-parsing checks.

Since the meta-attribute is not dependent on the existence of the
post-parsing directives, it is not neccessary to place it within the
conditional block.

Declare the variable and assign the attribute value at the top of the
function.