Adjust control focus when getting chapters
Upon the selection change of title uri combo, the chapters of the current
series automatically retrieved. When in focus and the user scrolls, the
combo changes the selected series, resulting in the chapters being
retrieved.

Fix this bug by removing focus of 'cbTitlUrl' when the 'GetChapters(Uri)'
method is executed. The check is exclusively for the title combo and
shouldn't affect other controls.

Once the chapters have been loaded into the chapter data grid, set focus
to it, allowing the user to scroll freely through the chapters.